fix: create checkingOffset function and add to ciphering func
