+++ src/validations/index.js
@@ -2,5 +2,6 @@ const isIncluded = require('./isIncluded');
 const isDuplicated = require('./isDuplicated');
 const testingFormat = require('./testingFormat');
 const findingFile = require('./findingFile');
+const checkingOffset = require('./checkingOffset');
 
-module.exports = { isIncluded, isDuplicated, testingFormat, findingFile };
+module.exports = { isIncluded, isDuplicated, testingFormat, findingFile, checkingOffset };
